requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==4.4.2
future==0.18.2
idna==3.1
pycparser==2.20
PySocks==1.7.1
pytest==6.2.2
python-dotenv==0.16.0
python-telegram-bot==13.4.1
requests==2.25.1
six==1.15.0
tornado==6.1
urllib3==1.26.4
